refactor: move health check use case to common
