+++ apps/op-app/package.json
@@ -15,6 +15,7 @@
     "fp-ts": "^2.16.6",
     "helmet": "^7.1.0",
     "i18n": "^0.15.1",
+    "io-fims-common": "workspace:^",
     "io-ts": "^2.2.21",
     "jose": "^5.4.0",
     "oidc-provider": "^8.4.6",
